Fix type conversion: UInt32 to Int32 for pid in ProcessListView.swift

+++ lara/views/ProcessListView.swift
@@ -173,7 +173,7 @@ struct ProcessListView: View {
                     }
                 }
                 procs.append(KernelProcess(
-                    pid: e.pid,
+                    pid: Int32(e.pid),
                     uid: e.uid,
                     gid: e.gid,
                     name: name,
